Fixes workspace cleanup failure when no '.active'
The workspace cleanup method calls the workspace delete method.
The workspace delete method used to delete the '.active' file from the
workspaces dir if the workspace that aimed to be deleted was active.
If the cleaned-up workspace was checked up before the cleanup and the
environment variable 'IR_WORKSPACE' was set, the activate method raise
an error and a new '.active' file wasn't created.
After that when try to delete an active workspace (defined in
IR_WORKSPACE), the delete operation failed becuase the '.active' file
was missing.
This patch re-activates only workspace that was active NOT by
'IR_WORKSPACE' env var.

RHOSINFRA-1191

+++ infrared/core/services/workspaces.py
@@ -293,7 +293,7 @@ class WorkspaceManager(object):
         if not self.has_workspace(name):
             raise exceptions.IRWorkspaceMissing(workspace=name)
         else:
-            if self.is_active(name):
+            if self.is_active(name) and os.path.exists(self.active_file):
                 os.remove(self.active_file)
             shutil.rmtree(os.path.join(self.workspace_dir, name))
 
@@ -436,10 +436,11 @@ class WorkspaceManager(object):
     def cleanup(self, name):
         """Removes all the files from the workspace folder"""
 
-        was_active = self.is_active(name)
+        checked_out = self.is_active(name) and os.environ.get(
+            ACTIVE_WORKSPACE_ENV_NAME) != name
         self.delete(name)
         self.create(name)
-        if was_active:
+        if checked_out:
             self.activate(name)
 
     def node_list(self, workspace_name=None):
